Modifica http request

+++ components/parse_endpoints/parse_endpoints.h
@@ -4,6 +4,6 @@
 #include "esp_err.h"
 
 esp_err_t confirmar_parse_dosificacion(char *confirmarDosificacionStr);
-char *crear_post_body(int consumo);
+char *crear_post_body(float consumo);
 
 #endif
